splie query sample to header and cpp
